gnhf 18: Extracted a shared useOneShotBurst hook in features/anim, collapsing the identical reduced-motion + active-toggle + fire→timeout burst logic duplicated across Confetti and Fireworks.
